requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
PyGithub==2.1.1
gunicorn==21.2.0
PyJWT==2.8.0
